+++ cpp/zimt/nsim.h
@@ -20,7 +20,7 @@
 #include <utility>
 #include <vector>
 
-#include "zimt/spectrogram.h"
+#include "zimt/zimtohrli.h"
 
 namespace zimtohrli {
